Move the classic parser to its own directory
