+++ link-grammar/parse/count.c
