+++ link-grammar/parse/extract-links.h
